Move HassioServiceInfo to init.py
