+++ homeassistant/components/ozw/config_flow.py
@@ -4,7 +4,7 @@ import logging
 import voluptuous as vol
 
 from homeassistant import config_entries
-from homeassistant.components.hassio.discovery import HassioServiceInfo
+from homeassistant.components.hassio import HassioServiceInfo
 from homeassistant.core import callback
 from homeassistant.data_entry_flow import AbortFlow, FlowResult
