+++ tests/components/hassio/test_discovery.py
@@ -2,7 +2,7 @@
 from http import HTTPStatus
 from unittest.mock import Mock, patch
 
-from homeassistant.components.hassio.discovery import HassioServiceInfo
+from homeassistant.components.hassio import HassioServiceInfo
 from homeassistant.components.hassio.handler import HassioAPIError
 from homeassistant.const import EVENT_HOMEASSISTANT_START
 from homeassistant.setup import async_setup_component
